# api/openapi.yaml
openapi: 3.1.0
info:
  title: AppHub control plane
  version: 1.0.0
  description: |
    Organization application and durable deployment API. This document is the wire
    contract for browser, CLI, SDK, and MCP adapter inputs and results. All API
    paths are relative to the operator-configured public origin. Production uses
    HTTPS; HTTP is allowed only on an explicitly configured literal loopback
    development origin. Origins and OAuth resources are never inferred from Host
    or forwarded headers.

    Users authenticate through an admitted Google or OIDC identity. Exact issuer
    and subject identify an account; equal email addresses never merge accounts.
    Eligibility, local disablement, admission, and operator-configured admin role
    are checked on each request. Ordinary users can access only their own
    applications and deployments. Admins can operate any application. Inaccessible
    object IDs return 404, before revealing scope failures on another owner's
    objects. Known accessible objects with insufficient scope return 403.

    API requests support a browser session OR an opaque AppHub OAuth bearer token
    for the exact resource {publicOrigin}/api. POST /api/v1/github/webhook is the
    exception: GitHub authenticates with X-Hub-Signature-256, and a session or
    bearer token is not a credential there. If Authorization is present, bearer
    authentication is selected exclusively; an invalid bearer never falls back to
...

...

...

...
servers:
  - url: /
...
security:
  - SessionCookie: []
  - OAuth: []
paths:
...
  /.well-known/oauth-authorization-server:
    get:
      operationId: getAuthorizationServerMetadata
      tags: [OAuth]
      summary: Discover the AppHub public-client OAuth authorization server
      security: []
      description: |
        Issuer is exactly publicOrigin. This is independent of upstream Google/OIDC
        login and workload identity. Only authorization-code and refresh grants are
        supported. Public clients authenticate with none and use S256 PKCE. No
        device grant, dynamic registration, introspection, or JWT/JWKS endpoint is
        provided. Only this exact discovery route is supported, not wildcard suffixes.
      responses:
        '200':
          description: Authorization server metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServerMetadata'
        '503':
          $ref: '#/components/responses/Unavailable'
  /.well-known/oauth-protected-resource/api:
    get:
      operationId: getAPIResourceMetadata
      tags: [OAuth]
      summary: Discover the exact API OAuth resource
      security: []
      responses:
        '200':
          description: resource is exactly {publicOrigin}/api.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectedResourceMetadata'
        '503':
          $ref: '#/components/responses/Unavailable'
  /.well-known/oauth-protected-resource/mcp:
    get:
      operationId: getMCPResourceMetadata
      tags: [OAuth]
      summary: Discover the exact remote MCP OAuth resource
      security: []
      responses:
        '200':
          description: resource is exactly {publicOrigin}/mcp; an API token is not valid here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectedResourceMetadata'
        '503':
          $ref: '#/components/responses/Unavailable'
  /oauth/authorize:
    get:
      operationId: beginAuthorization
      tags: [OAuth]
      summary: Validate an OAuth request and begin browser-bound consent
      security: []
      description: |
        Validates all parameters before redirecting. Unauthenticated browsers use
        the normal provider picker and resume this request after login. Stores a
        five-minute consent transaction binding session, user, client, exact
        redirect, state, resource, exact scopes, and PKCE, then opens /authorize.
        The built-in public client apphub-cli uses http://127.0.0.1:<ephemeral-port>/callback
        with no userinfo, query, or fragment. Operator-registered remote clients
        require exact HTTPS redirects. Remote client IDs may be HTTPS Client ID
        Metadata Document URLs; bounded, nonredirecting, public-IP-pinned fetching
        verifies exact client_id and public none authentication. No private,
        link-local, metadata, redirect, or environment-proxy fetch is permitted.
        Neither wildcard callbacks nor anonymous dynamic registration are supported.
        Never redirect an error until the client and redirect URI have been validated.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            const: code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: resource
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/OAuthResource'
        - name: scope
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ScopeString'
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{43}$'
            description: Unpadded base64url SHA-256 of a PKCE verifier.
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            const: S256
      responses:
        '302':
          description: |
            Redirect to sign-in, local consent UI, or a validated OAuth error
            callback. Callback responses include the original state and exact iss.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/OAuthBadRequest'
        '429':
          $ref: '#/components/responses/OAuthRateLimited'
        '503':
          $ref: '#/components/responses/OAuthUnavailable'
    post:
      operationId: decideConsent
      tags: [OAuth]
      summary: Atomically consume bound consent and approve or deny
      security:
        - SessionCookie: []
      description: |
        Requires the exact bound browser session, Origin, and CSRF token. Accepts
        only transactionId and literal action. Redirect, resource, scope, client,
        state, and PKCE come only from the stored transaction. A transaction is
        single-use; denial produces access_denied and no code. Approval produces
        a single-use five-minute authorization code. Success and denial callback
        URLs both include the correlated state and exact issuer in iss (RFC 9207).
        Codes and transactions are never exposed in logs or diagnostic bodies.
        Returns a JSON redirectUrl for explicit browser navigation; it does not
        make fetch follow an external redirect with browser credentials.
      parameters:
        - $ref: '#/components/parameters/RequiredOrigin'
        - $ref: '#/components/parameters/RequiredCSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentDecisionInput'
      responses:
        '200':
          description: Bound callback URL with code or error, state, and iss; navigate the browser explicitly.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentDecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
        '503':
          $ref: '#/components/responses/Unavailable'
  /oauth/token:
    post:
      operationId: exchangeToken
      tags: [OAuth]
      summary: Exchange a bound code or rotate a refresh token
      security: []
      description: |
        Public clients use client_id and no client secret. Code exchange requires
        exact client, redirect, resource, and a matching S256 verifier, atomically
        consumes the code, and creates an active OAuth family and tokens. Access
        tokens are opaque, valid for 15 minutes, and bound to exactly one resource.
        OAuth families have an absolute 30-day lifetime that refresh never extends.
        Refresh checks family state, expiry, client, resource, and current user
        eligibility, then atomically consumes and rotates the refresh token.
        Consumed refresh token replay revokes the entire family; revocation cannot
        be undone by a concurrent refresh. Clients must reauthenticate after an
        uncertain refresh transport outcome rather than replay a possibly consumed
        token. Credential failures use generic invalid_grant, not provider details;
        storage failures remain unavailable. Bearer authentication checks active
        token, family, audience, and current eligibility on every request.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
                - $ref: '#/components/schemas/AuthorizationCodeExchange'
                - $ref: '#/components/schemas/RefreshTokenExchange'
      responses:
        '200':
          description: New access token and rotated refresh token; never cache.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            Pragma:
              description: Legacy cache prevention.
              schema:
                type: string
                const: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/OAuthBadRequest'
        '413':
          $ref: '#/components/responses/TooLarge'
        '429':
          $ref: '#/components/responses/OAuthRateLimited'
        '503':
          $ref: '#/components/responses/OAuthUnavailable'
  /oauth/revoke:
    post:
      operationId: revokeOAuthToken
      tags: [OAuth]
      summary: Durably revoke the OAuth family associated with a token
      security: []
      description: |
        Public-client revocation with client_id and the token, not browser cookies.
        Unknown or already revoked tokens return the same success without revealing
        validity. A known token must match the client. Family revocation invalidates
        all associated access/refresh tokens and persists across concurrent refresh.
        Storage failure is unavailable, not successful revocation.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RevokeTokenInput'
      responses:
        '200':
          description: Revocation processed; empty response, including for unknown tokens.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/OAuthBadRequest'
        '413':
          $ref: '#/components/responses/TooLarge'
        '429':
          $ref: '#/components/responses/OAuthRateLimited'
        '503':
          $ref: '#/components/responses/OAuthUnavailable'
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-apphub_session
      description: |
        Host-only Secure HttpOnly SameSite=Lax session, Path=/, no Domain. Explicit
        loopback HTTP uses apphub_session instead. Cookie mutations require exact
        Origin and X-CSRF-Token. Never accepted as an OAuth token or at /mcp.
    OAuth:
      type: oauth2
      description: |
        Opaque AppHub bearer tokens only. API operations require resource exactly
        {publicOrigin}/api; MCP requires {publicOrigin}/mcp. Send the resource in
        both authorization and code exchange. Scope names are exact set members,
        not prefixes. Every token must have at least one known scope. The empty
        requirement list on user/session operations means any nonempty granted
        scope is sufficient, not an unscoped token. Browser sessions do not need
        OAuth scopes. Invalid Authorization never falls back to a cookie.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            applications:read: Read accessible applications and safe target options.
            applications:write: Create and edit accessible applications.
            deployments:read: Read accessible deployment operations and history.
            deployments:write: Submit deployments; resolve interrupted work only if also admin.
  parameters:
...
    RequiredOrigin:
      name: Origin
      in: header
      required: true
      description: Must exactly equal configured publicOrigin; missing and foreign origins are rejected.
      schema:
        type: string
        format: uri
    RequiredCSRF:
      name: X-CSRF-Token
      in: header
      required: true
      description: Session-bound csrfToken from users/me; checked together with the exact Origin.
      schema:
        type: string
        minLength: 1
  headers:
...
    Location:
      description: Redirect target or durable deployment status URI, depending on the operation.
      schema:
        type: string
        format: uri-reference
    NoStore:
      description: Sensitive authentication/session/token responses are never cached.
      schema:
        type: string
        const: no-store
...
  responses:
    BadRequest:
      description: Malformed JSON, unknown field, invalid query/header, or missing required precondition.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Missing, expired, revoked, ineligible, or invalid authentication; no cookie fallback after invalid bearer.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/Authenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient exact scope or role, denied admission, or missing/invalid browser Origin or CSRF token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource missing or inaccessible to this principal; cross-owner IDs do not reveal existence.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: |
        Revision mismatch, idempotency_conflict, active deployment, unresolved
        interrupted lock, hostname reservation conflict, or invalid state transition.
        No cloud mutation is initiated by a rejected request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooLarge:
      description: JSON request exceeds 128 KiB or serialized application specification exceeds 64 KiB; never truncated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: Required storage, authentication provider, target configuration, or worker readiness unavailable; retryable, never an empty success.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    OAuthBadRequest:
      description: Invalid OAuth request or grant; credential failures use generic invalid_grant.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAuthError'
    OAuthRateLimited:
      description: OAuth request rate limited.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAuthError'
    OAuthUnavailable:
      description: OAuth dependency unavailable; not an invalid credential or successful revocation.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAuthError'
  schemas:
...
    ConsentDecisionInput:
      type: object
      additionalProperties: false
      required: [transactionId, action]
      properties:
        transactionId:
          type: string
          minLength: 1
        action:
          type: string
          enum: [approve, deny]
    ConsentDecisionResponse:
      type: object
      additionalProperties: false
      required: [redirectUrl]
      properties:
        redirectUrl:
          type: string
          format: uri
          description: Exact transaction-bound callback plus code or error, correlated state, and issuer; explicitly navigate the browser to this URL.
...
    ScopeString:
      type: string
      minLength: 1
      pattern: '^(applications:read|applications:write|deployments:read|deployments:write)( (applications:read|applications:write|deployments:read|deployments:write))*$'
      description: Nonempty space-delimited set of exact supported scopes; unknown scopes are rejected.
    OAuthResource:
      type: string
      format: uri
      description: Exactly {publicOrigin}/api, {publicOrigin}/mcp, or an enabled hosted application's https://{applicationHost}/mcp audience; one resource per token/family, never a prefix.
    AuthorizationServerMetadata:
      type: object
      required: [issuer, authorization_endpoint, token_endpoint, revocation_endpoint, response_types_supported, grant_types_supported, code_challenge_methods_supported, token_endpoint_auth_methods_supported, scopes_supported, authorization_response_iss_parameter_supported]
      properties:
        issuer:
          type: string
          format: uri
          description: Exactly configured publicOrigin, independent of request headers.
        authorization_endpoint:
          type: string
          format: uri
        token_endpoint:
          type: string
          format: uri
        revocation_endpoint:
          type: string
          format: uri
        response_types_supported:
          type: array
          items:
            type: string
            const: code
          const: [code]
        grant_types_supported:
          type: array
          items:
            type: string
            enum: [authorization_code, refresh_token]
          minItems: 2
          maxItems: 2
          uniqueItems: true
        code_challenge_methods_supported:
          type: array
          items:
            type: string
            const: S256
          const: [S256]
        token_endpoint_auth_methods_supported:
          type: array
          items:
            type: string
            const: none
          const: [none]
        revocation_endpoint_auth_methods_supported:
          type: array
          items:
            type: string
            const: none
          const: [none]
        scopes_supported:
          type: array
          minItems: 4
          maxItems: 4
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Scope'
        authorization_response_iss_parameter_supported:
          type: boolean
          const: true
        client_id_metadata_document_supported:
          type: boolean
          const: true
    ProtectedResourceMetadata:
      type: object
      required: [resource, authorization_servers, scopes_supported, bearer_methods_supported]
      properties:
        resource:
          $ref: '#/components/schemas/OAuthResource'
        authorization_servers:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: string
            format: uri
          description: Exactly one entry, the configured publicOrigin issuer.
        scopes_supported:
          type: array
          minItems: 4
          maxItems: 4
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Scope'
        bearer_methods_supported:
          type: array
          items:
            type: string
            const: header
          const: [header]
    AuthorizationCodeExchange:
      type: object
      additionalProperties: false
      required: [grant_type, code, client_id, redirect_uri, resource, code_verifier]
      properties:
        grant_type:
          type: string
          const: authorization_code
        code:
          type: string
          minLength: 1
          description: Single-use code; expires five minutes after approval.
        client_id:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          format: uri
          description: Exact redirect_uri bound at authorization.
        resource:
          $ref: '#/components/schemas/OAuthResource'
        code_verifier:
          type: string
          minLength: 43
          maxLength: 128
          pattern: '^[A-Za-z0-9._~-]{43,128}$'
    RefreshTokenExchange:
      type: object
      additionalProperties: false
      required: [grant_type, refresh_token, client_id, resource]
      properties:
        grant_type:
          type: string
          const: refresh_token
        refresh_token:
          type: string
          minLength: 1
          description: Current unconsumed refresh token; replay revokes the family.
        client_id:
          type: string
          minLength: 1
        resource:
          $ref: '#/components/schemas/OAuthResource'
...